# docs/openapi-core.yaml
openapi: 3.1.0
info:
  title: Java API Dishwasher
  description: Smart dishwasher challenge
  version: 0.0.4
  contact:
    name: Dishwasher
servers:
  - url: http://localhost:4000
...
paths:
  /programs:
    post:
      tags:
        - programs
      operationId: startProgram
      summary: POST Program to start (this should not be available if a program is already running)
      description: Start program
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartProgram'
      responses:
        '201':
          description: Program started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Program'
    get:
      tags:
        - programs
      summary: GET available programs to run
      description: GET available programs to run
      operationId: getPrograms
      responses:
        '200':
          description: Get all programs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllPrograms'
...
components:
  schemas:
    StartProgram:
      type: object
      required:
    Program:
      type: object
    AllPrograms:
      type: array
      items:
        $ref: '#/components/schemas/Program'